\name{genUSStateSites}
\alias{genUSStateSites}
\title{ Generating Random Sites in a US State }
\description{
  A completely spatial random set of point is generated for a US state.
}
\usage{
genUSStateSites(state, nsites)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{state}{ the name of a US state. }
  \item{nsites}{ the number of sites needed. }
}
\value{
  A matrix of longitude and latitude....
}
\seealso{  \code{\link{genUSStateSites}} }
\examples{

}
\keyword{ datagen }
\keyword{ utilities }
